Try using different colors to highlight power zones
This is a first attempt at using colors to illustrate to the rider what
power zone they were in, my idea was that when you're in zone 1 the
power bar would fill up as the power increased, when you get to zone 2
the bar would have a 'grey' background from the zone one, but then start
the blue from the zone two would start to fill the bar from the left, so
not only can you see what zone you are in, but also where in that zone
you are as a percentage.

This didn't come out as clearly as I would have hoped so may need to
tweak things to change up the idea entirely to achieve the same goal,
perhaps we could fill up the same bar but with multiple colors as the
rider gets to different zones.

... needs more thought