Refactor ArcField to save memory
Remove the dictionaries that were used to store the zones for power/hr/speed
and use arrays instead. Only create those zone objects for the field that we're
displaying to cut down more on memory.

Also removed some comments about using shared objects as I don't think this is
realistic as it would create a large amount of global memory that can't be GCd.

This change takes peak memory down from 24.0KB to 21.7KB