Add layout class and increase font size
The class is temporary, but it helps clean up the hard coded UI values
which would break with devices of different screen resolutions. Now
everything is scaled based on the device using the L class. Later I'd
like to remove this and just use getWidth() and getHeight() directly to
save on memory as this change did increase peak memory usage by around
0.4KB