Support metric and imperial units depending on device preference
If the user has imperial units, we now display the correct values and units.
This means that our raw data is stored in whatever format comes from the garmin
API, then we multiply when displaying the values. This means that I had to
change some internal calculations as they were using speed and assuming km/h.